Cleanup and move task registration logic into separate files in relevant packages
